Bug while deleting image rectified

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -151,9 +151,26 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+
+        Image image = imageService.getImage(imageId);
+        List<Tag> tags = image.getTags();
+        // setting model attributes first and then deleting later if user matches with sessionUser
+        model.addAttribute("image", image);
+        model.addAttribute("tags",tags);
+        String errorMessage = "Error";
+        User user = imageService.getImage(imageId).getUser();
+        User sessionUser = (User)session.getAttribute("loggeduser");
+        if (user != null && user.getId().equals(sessionUser.getId())) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else {
+            model.addAttribute("deleteError", errorMessage);
+            return "images/image";
+        }
+
+
     }
